LP/60 Data Correction Modal change to Component
Deleted UseFeedbackModal
Added modal component to ProfileInfo
