+++ src/Web/src/components/ProfileComponents/ProfileInfo.js
@@ -1,24 +1,10 @@
 import React, { useState } from 'react';
 import { Row, Col,  Button, Card, Modal, ModalHeader, ModalBody, ModalFooter, Label, Input, Form, FormGroup } from 'reactstrap';
 import { PersonIcon, GeoIcon, PhoneIcon, MailIcon, EducationIconNavy, RibbonIcon, CalendarIcon, IdIcon } from '../Icons';
-import UseFeedbackModal from './UseFeedbackModal';
+import DataCorrectionModal from '../DataCorrectionComponent/DataCorrectionModal';
 
 const ProfileInfo = (props) => {
     const { data } = props;
-    const {sendFeedback, bind} = UseFeedbackModal();
-    
-    const [modal, setModal] = useState(false);
-    const toggle = () => setModal(!modal);
-
-    const handleOnSubmit = (e) => {
-        e.preventDefault();
-        const data = {"sui": e.currentTarget[0].value,
-            "fullname": e.currentTarget[1].value,
-            "email": e.currentTarget[2].value}
-        sendFeedback(data);
-
-        setModal(false);
-    };
 
     return (
         <Card body>
@@ -65,46 +51,12 @@ const ProfileInfo = (props) => {
                                 <td></td>
                                 <td></td>
                                 <td>
-                                 <a href="#" className="modal-link" onClick={toggle}>Request Data Correction</a>
+                                    <DataCorrectionModal data={data}></DataCorrectionModal>
                                 </td>
                             </tr>
                         </tbody>
                     </table>
-                    <Modal isOpen={modal} toggle={toggle} fullscreen="sm">
-                        <ModalHeader toggle={toggle}>Data Correction Request Form</ModalHeader>
-                        <ModalBody>
-                            <Form id="feedback-form" onSubmit={e => handleOnSubmit(e)}>
-                                <Input type="hidden" name="staffUniqueId" id="staffUniqueId" value={data.staffUniqueId}></Input>
-                                <Input type="hidden" name="fullName" id="fullName" value={data.fullName}></Input>
-                                <Input type="hidden" name="email" id="email" value={data.email}></Input>
-
-                                <div className="mb-2 mr-sm-2 mb-sm-0">
-                                    <Label for="lblsubject" className="label-feedback mr-sm-2">ID: </Label>
-                                    <Label for="lblsubject" className="mr-sm-2">{data.staffUniqueId}</Label>
-                                </div>
-                                <div className="mb-2 mr-sm-2 mb-sm-0">
-                                    <Label for="staffsubject" className="label-feedback mr-sm-2">Name: </Label>
-                                    <Label for="staffsubject" className="mr-sm-2">{data.fullName}</Label>
-                                </div>
-                                <div className="mb-2 mr-sm-2 mb-sm-0">
-                                    <Label for="emailsubject" className="label-feedback mr-sm-2">Email: </Label>
-                                    <Label for="emailsubject" className="mr-sm-2">{data.email}</Label>
-                                </div>
-                                <FormGroup className="mb-2 mr-sm-2 mb-sm-0">
-                                    <Label for="subject" className="label-feedback mr-sm-2">Subject</Label>
-                                    <Input type="text" placeholder="" name="subject" id="subject" value="Data Correction Request" {...bind}/>
-                                </FormGroup>
-                                <FormGroup className="mb-2 mr-sm-2 mb-sm-0">
-                                    <Label for="messagescontent" className="label-feedback mr-sm-2">Description</Label>
-                                    <Input type="textarea" name="messagescontent" id="messagecontent"  {...bind}/>
-                                </FormGroup>
-                            </Form>
-                        </ModalBody>
-                        <ModalFooter>
-                            <Button color="primary" form="feedback-form">Submit</Button>{' '}
-                            <Button color="secondary" onClick={toggle}>Cancel</Button>
-                        </ModalFooter>
-                    </Modal>
+                    
                 </Col>
             </Row>
         </Card>
